src/main/java/com/odde/bbuddy/budget/view/PresentableMonthlyBudgetAmount.java
package com.odde.bbuddy.budget.view;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.context.annotation.PropertySource;
import org.springframework.context.annotation.Scope;
import org.springframework.stereotype.Component;
import org.springframework.web.servlet.ModelAndView;

import static com.odde.bbuddy.common.controller.Urls.MONTHLYBUDGET_TOTALAMOUNT;
import static com.odde.bbuddy.common.view.MessageSources.RESULT_MESSAGES_FULL_NAME;
import static java.lang.String.format;
import static org.springframework.context.annotation.ScopedProxyMode.TARGET_CLASS;

@Component
@Scope(value = "request", proxyMode = TARGET_CLASS)
@PropertySource(RESULT_MESSAGES_FULL_NAME)
public class PresentableMonthlyBudgetAmount extends ModelAndView {

    @Value("${monthlybudget.totalamount.amount}")
    public String message;

    public PresentableMonthlyBudgetAmount() {
        setViewName(MONTHLYBUDGET_TOTALAMOUNT);
    }

    public ModelAndView display(long amount) {
        addObject("amount", format(message, amount));
        return this;
    }
}
